requirements.txt
pandas>=2.2
numpy>=2.0
matplotlib>=3.8
requests>=2.32
python-dotenv>=1.0
